add support for methods using contexts
