+++ .gitignore
@@ -1,3 +1,4 @@
 _*
 *.out
 *~
+.idea
